IO-429: Test case for IOUtils.toByteArray() with long InputStream
Add a test case testToByteArray_InputStreamTooLong() to IOUtilsTestCase
to test how InputStreams with more than Integer.MAX_VALUE bytes are
processed by IOUtils.toByteArray().

The test case requires huge amounts of memory, thus memory limit is
increased to 4223M (the smallest value with which the new test case does
not fail). The limit was introduced in
https://issues.apache.org/jira/browse/IO-161, but the reasons for the
exact value are not detailed.